// src/Models/iMovie.ts
export interface IMovie {
    adult: boolean
    backdrop_path: string
    genre_ids: number[]
    id: number
    original_language: string
    original_title: string
    overview: string
    popularity: any
    poster_path: string
    release_date: string
    title: string
    video: false
    vote_average: any
    vote_count: any
}